move schemas into dir
